validation: move CastCall helper from superchain to validation package
