+++ validation/utils_test.go
@@ -2,35 +2,10 @@ package validation
 
 import (
 	"fmt"
-	"math/big"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-// isBigIntWithinBounds returns true if actual is within bounds, where the bounds are [lower bound, upper bound] and are inclusive.
-var isBigIntWithinBounds = func(actual *big.Int, bounds [2]*big.Int) bool {
-	if (bounds[1].Cmp(bounds[0])) < 0 {
-		panic("bounds are in wrong order")
-	}
-	return (actual.Cmp(bounds[0]) >= 0 && actual.Cmp(bounds[1]) <= 0)
-}
-
-// isIntWithinBounds returns true if actual is within bounds, where the bounds are [lower bound, upper bound] and are inclusive.
-func isIntWithinBounds[T uint32 | uint64](actual T, bounds [2]T) bool {
-	if bounds[1] < bounds[0] {
-		panic("bounds are in wrong order")
-	}
-	return (actual >= bounds[0] && actual <= bounds[1])
-}
-
-// assertInBounds fails the test (but not immediately) if the passed param is outside of the passed bounds.
-func assertIntInBounds[T uint32 | uint64](t *testing.T, name string, got T, want [2]T) {
-	assert.True(t,
-		isIntWithinBounds(got, want),
-		fmt.Sprintf("Incorrect %s, %d is not within bounds %d", name, got, want))
-}
 
 func TestIsIntWithinBounds(t *testing.T) {
 	tt := []struct {
